+++ packages/page-navigation/src/GoToLastPageButton.tsx
@@ -10,11 +10,11 @@ import * as React from 'react';
 import { Button, LocalizationContext, Position, Tooltip } from '@react-pdf-viewer/core';
 
 import DownArrowIcon from './DownArrowIcon';
-import { RenderGoToLastPageProps } from './GoToLastPage';
+import { RenderGoToPageProps } from './types/index';
 
 const TOOLTIP_OFFSET = { left: 0, top: 8 };
 
-const GoToLastPageButton: React.FC<RenderGoToLastPageProps> = ({ isDisabled, onClick }) => {
+const GoToLastPageButton: React.FC<RenderGoToPageProps> = ({ isDisabled, onClick }) => {
     const l10n = React.useContext(LocalizationContext);
     const label = l10n && l10n.pageNavigation ? l10n.pageNavigation.goToLastPage : 'Last page';