+++ packages/page-navigation/src/GoToNextPageMenuItem.tsx
@@ -10,9 +10,9 @@ import * as React from 'react';
 import { LocalizationContext, MenuItem } from '@react-pdf-viewer/core';
 
 import NextIcon from './NextIcon';
-import { RenderGoToNextPageProps } from './GoToNextPage';
+import { RenderGoToPageProps } from './types/index';
 
-const GoToNextPageMenuItem: React.FC<RenderGoToNextPageProps> = ({ isDisabled, onClick }) => {
+const GoToNextPageMenuItem: React.FC<RenderGoToPageProps> = ({ isDisabled, onClick }) => {
     const l10n = React.useContext(LocalizationContext);
     const label = l10n && l10n.pageNavigation ? l10n.pageNavigation.goToNextPage : 'Next page';