+++ packages/page-navigation/src/GoToPreviousPageButton.tsx
@@ -9,12 +9,12 @@
 import * as React from 'react';
 import { Button, LocalizationContext, Position, Tooltip } from '@react-pdf-viewer/core';
 
-import { RenderGoToPreviousPageProps } from './GoToPreviousPage';
 import PreviousIcon from './PreviousIcon';
+import { RenderGoToPageProps } from './types/index';
 
 const TOOLTIP_OFFSET = { left: 0, top: 8 };
 
-const GoToPreviousPageButton: React.FC<RenderGoToPreviousPageProps> = ({ isDisabled, onClick }) => {
+const GoToPreviousPageButton: React.FC<RenderGoToPageProps> = ({ isDisabled, onClick }) => {
     const l10n = React.useContext(LocalizationContext);
     const label = l10n && l10n.pageNavigation ? l10n.pageNavigation.goToPreviousPage : 'Previous page';